+++ src/exactly_lib/definitions/primitives/line_matcher.py
@@ -1,5 +1,2 @@
-from exactly_lib.definitions.primitives import str_matcher
-
-REGEX_MATCHER_NAME = str_matcher.MATCH_REGEX_OR_GLOB_PATTERN_CHECK_ARGUMENT
 LINE_NUMBER_MATCHER_NAME = 'line-num'
 CONTENTS_MATCHER_NAME = 'contents'