+++ src/exactly_lib/test_case_utils/line_matcher/parse_line_matcher.py
@@ -7,7 +7,7 @@ from exactly_lib.definitions.entity import types
 from exactly_lib.definitions.primitives import line_matcher
 from exactly_lib.test_case_utils.expression import grammar, parser as ep
 from exactly_lib.test_case_utils.expression.parser import GrammarParsers
-from exactly_lib.test_case_utils.line_matcher.impl import matches_regex, line_number
+from exactly_lib.test_case_utils.line_matcher.impl import line_number
 from exactly_lib.test_case_utils.line_matcher.impl.contents import parse as contents_parse, doc as contents_doc
 from exactly_lib.test_case_utils.matcher import standard_expression_grammar
 from exactly_lib.type_system.logic.line_matcher import FIRST_LINE_NUMBER, LineMatcherSdv
@@ -23,15 +23,10 @@ def parsers(must_be_on_current_line: bool = False) -> GrammarParsers[LineMatcher
 
 
 _TP = TextParser({
-    'REG_EX': syntax_elements.REGEX_SYNTAX_ELEMENT.singular_name,
     'INTEGER_MATCHER': syntax_elements.INTEGER_MATCHER_SYNTAX_ELEMENT.singular_name,
     'FIRST_LINE_NUMBER': FIRST_LINE_NUMBER,
     'MODEL': matcher_model.LINE_MATCHER_MODEL,
 })
-
-_REGEX_MATCHER_SED_DESCRIPTION = """\
-Matches {MODEL:s} that contain a string that matches {REG_EX}.
-"""
 
 _LINE_NUMBER_MATCHER_SED_DESCRIPTION = """\
 Matches {MODEL:s} who's line number matches {INTEGER_MATCHER}.
@@ -39,22 +34,6 @@ Matches {MODEL:s} who's line number matches {INTEGER_MATCHER}.
 
 Line numbers start at {FIRST_LINE_NUMBER}.
 """
-
-
-class _RegexSyntaxDescription(grammar.PrimitiveDescriptionWithNameAsInitialSyntaxToken):
-    @property
-    def argument_usage_list(self) -> Sequence[a.ArgumentUsage]:
-        return [
-            syntax_elements.REGEX_SYNTAX_ELEMENT.single_mandatory,
-        ]
-
-    @property
-    def description_rest(self) -> Sequence[ParagraphItem]:
-        return _TP.fnap(_REGEX_MATCHER_SED_DESCRIPTION)
-
-    @property
-    def see_also_targets(self) -> Sequence[SeeAlsoTarget]:
-        return [syntax_elements.REGEX_SYNTAX_ELEMENT.cross_reference_target]
 
 
 class _LineNumberSyntaxDescription(grammar.PrimitiveDescriptionWithNameAsInitialSyntaxToken):
@@ -85,19 +64,14 @@ GRAMMAR = standard_expression_grammar.new_grammar(
     value_type=ValueType.LINE_MATCHER,
     simple_expressions=(
         NameAndValue(
-            line_matcher.LINE_NUMBER_MATCHER_NAME,
-            grammar.Primitive(line_number.parse_line_number,
-                              _LineNumberSyntaxDescription())
-        ),
-        NameAndValue(
             line_matcher.CONTENTS_MATCHER_NAME,
             grammar.Primitive(contents_parse.PARSER.parse,
                               contents_doc.SyntaxDescription())
         ),
         NameAndValue(
-            line_matcher.REGEX_MATCHER_NAME,
-            grammar.Primitive(matches_regex.parse,
-                              _RegexSyntaxDescription())
+            line_matcher.LINE_NUMBER_MATCHER_NAME,
+            grammar.Primitive(line_number.parse_line_number,
+                              _LineNumberSyntaxDescription())
         ),
     ),
 )