+++ test/exactly_lib_test/instructions/assert_/contents_of_file/relativity_option_for_actual_file/any_line_matches.py
@@ -15,7 +15,8 @@ from exactly_lib_test.instructions.assert_.test_resources.instruction_check impo
 from exactly_lib_test.tcfs.test_resources.sub_dir_of_sds_act import \
     MK_SUB_DIR_OF_ACT_AND_MAKE_IT_CURRENT_DIRECTORY
 from exactly_lib_test.test_case.test_resources.arrangements import ArrangementPostAct
-from exactly_lib_test.test_case_utils.line_matcher.test_resources.argument_syntax import syntax_for_regex_matcher
+from exactly_lib_test.test_case_utils.line_matcher.test_resources import arguments_building as lm_args
+from exactly_lib_test.test_case_utils.string_matcher.test_resources import arguments_building2 as sm_args
 from exactly_lib_test.test_case_utils.string_matcher.test_resources.arguments_building import \
     LineMatchesAssertionArgumentsConstructor
 from exactly_lib_test.test_resources.files.file_structure import DirContents, File, Dir
@@ -110,5 +111,7 @@ class _ContentsContainsALineThatMatches(TestWithConfigurationAndRelativityOption
 
 
 def exists_line_matches_reg_ex(regex: str) -> str:
-    return str(LineMatchesAssertionArgumentsConstructor(Quantifier.EXISTS,
-                                                        syntax_for_regex_matcher(regex)))
+    return str(LineMatchesAssertionArgumentsConstructor(
+        Quantifier.EXISTS,
+        lm_args.Contents(sm_args.Matches(regex)).as_str)
+    )