+++ test/exactly_lib_test/test_case_utils/line_matcher/z_package_suite.py
@@ -1,6 +1,6 @@
 import unittest
 
-from exactly_lib_test.test_case_utils.line_matcher import line_number, regex, std_expr
+from exactly_lib_test.test_case_utils.line_matcher import line_number, std_expr
 from exactly_lib_test.test_case_utils.line_matcher import model_construction
 from exactly_lib_test.test_case_utils.line_matcher import parse_line_matcher
 from exactly_lib_test.test_case_utils.line_matcher.contents import z_package_suite as contents
@@ -12,7 +12,6 @@ def suite() -> unittest.TestSuite:
         std_expr.suite(),
         line_number.suite(),
         contents.suite(),
-        regex.suite(),
         parse_line_matcher.suite(),
     ])